feat: add specific timeout after first request
also: rework option parsing, remove clippy warning, add RFC error code

+++ src/server.rs
@@ -339,7 +339,7 @@ fn parse_options(
                     *value = size as usize;
                     worker_options.transfer_size = size as usize;
                 }
-                RequestType::Write => worker_options.transfer_size = *value as usize,
+                RequestType::Write => worker_options.transfer_size = *value,
             },
             OptionType::Timeout => {
                 if *value == 0 {
